move tx_fee to helper
