+++ storage/blockchain/src/free.rs
@@ -1,6 +1,5 @@
 //! General free functions (related to the database).
 
-use monero_serai::transaction::{Input, Transaction};
 //---------------------------------------------------------------------------------------------------- Import
 use cuprate_database::{ConcreteEnv, Env, EnvInner, InitError, RuntimeError, TxRw};
 
@@ -60,37 +59,6 @@ pub fn open(config: Config) -> Result<ConcreteEnv, InitError> {
     }
 
     Ok(env)
-}
-
-//---------------------------------------------------------------------------------------------------- Tx Fee
-/// Calculates the fee of the [`Transaction`].
-///
-/// # Panics
-/// This will panic if the inputs overflow or the transaction outputs too much.
-pub(crate) fn tx_fee(tx: &Transaction) -> u64 {
-    let mut fee = 0_u64;
-
-    match &tx {
-        Transaction::V1 { prefix, .. } => {
-            for input in &prefix.inputs {
-                match input {
-                    Input::Gen(_) => return 0,
-                    Input::ToKey { amount, .. } => {
-                        fee = fee.checked_add(amount.unwrap_or(0)).unwrap();
-                    }
-                }
-            }
-
-            for output in &prefix.outputs {
-                fee.checked_sub(output.amount.unwrap_or(0)).unwrap();
-            }
-        }
-        Transaction::V2 { proofs, .. } => {
-            fee = proofs.as_ref().unwrap().base.fee;
-        }
-    };
-
-    fee
 }
 
 //---------------------------------------------------------------------------------------------------- Tests
